### 2019-09-02 #49 @whedon
[PRE REVIEW]: Basic Threading Examples in JuliaLang v1.3
**Submitting author:** @vtjnash (<a href="http://orcid.org/0000-0002-8991-0523">Jameson Nash</a>)
**Repository:** <a href="https://github.com/vtjnash/JuliaCon2019Threading" target ="_blank">https://github.com/vtjnash/JuliaCon2019Threading</a>

Thanks for submitting your paper to JuliaCon @vtjnash. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@vtjnash if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
