### 2019-09-02 #49 @vchuravy
@vtjnash please follow the instructions at https://proceedings.juliacon.org/guide/authors/ and setup a Github repository with the paper.
