+++ paper.tex
@@ -1,0 +1,592 @@
+
+% Based on the JuliaCon proceedings template
+% Layout inspired by the International Journal of Computer Applications template
+\documentclass{juliacon}
+\usepackage{hgblistings}
+\usepackage[many]{tcolorbox}
+\usepackage[framemethod=tikz]{mdframed}
+\lstMakeShortInline[language=Julia,columns=fixed]|
+\setcounter{page}{1}
+
+\newtcolorbox{calloutbox}[1][]{
+  breakable,
+  freelance,
+  title=#1,
+  colback=lightgray,
+  colbacktitle=lightgray,
+  coltitle=black,
+  fonttitle=\bfseries,
+  bottomrule=0pt,
+  boxrule=0pt,
+  colframe=lightgray,
+  overlay unbroken and first={
+  \draw[gray!75!black,line width=3pt]
+    ([xshift=5pt]frame.north west) --
+    (frame.north west) --
+    (frame.south west);
+  \draw[gray!75!black,line width=3pt]
+    ([xshift=-5pt]frame.north east) --
+    (frame.north east) --
+    (frame.south east);
+  },
+  overlay unbroken app={
+  \draw[gray!75!black,line width=3pt,line cap=rect]
+    (frame.south west) --
+    ([xshift=5pt]frame.south west);
+  \draw[gray!75!black,line width=3pt,line cap=rect]
+    (frame.south east) --
+    ([xshift=-5pt]frame.south east);
+  },
+  overlay middle and last={
+  \draw[gray!75!black,line width=3pt]
+    (frame.north west) --
+    (frame.south west);
+  \draw[gray!75!black,line width=3pt]
+    (frame.north east) --
+    (frame.south east);
+  },
+  overlay last app={
+  \draw[gray!75!black,line width=3pt,line cap=rect]
+    (frame.south west) --
+    ([xshift=5pt]frame.south west);
+  \draw[gray!75!black,line width=3pt,line cap=rect]
+    (frame.south east) --
+    ([xshift=-5pt]frame.south east);
+  },
+}
+
+
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+A major distinguishing point of any programming language is how it deals with concurrency. Of course, I'm really interested here in parallelism, and specifically how to get the best throughput out of my application. But we know that actually taking advantage of all cores correctly and efficiently is really hard as a programmer. Here we take a look at how JuliaLang prepares to unleash the full power of your modern CPU's many cores.
+
+Hopefully we won't drive any users mad with frustration as they try to write code to utilize threads. Towards that end, we'll look at how JuliaLang looks to provide a range of modern primitives that are designed to automatically compose effectively, and some of the trade-offs we make to try to simplify the mental model for the user. We'll also take a look at where we think that this development should go towards next.
+
+
+\end{abstract}
+
+\section{Introduction}
+\label{subsub:intro}
+
+A design principle for JuliaLang is to make common tasks easy and difficult tasks possible. This has extended from automatic memory management vs manual memory reuse, to having one type of method with dispatch, to optional type-inference for performance. And now, this includes concurrency. By extending design work that has been present in the language for many years to provide parallelism, we developed a means to add parallelism that preserved the (relative) simplicity of single-threaded for existing code, while allowing new code to benefit from the possibility of threading. This work has been inspired by parallel programming systems like Cilk, Intel Threading Building Blocks (TBB) and Go.
+
[…]
