+++ header.tex
@@ -1,0 +1,9 @@
+\title{Basic Threading Examples in JuliaLang v1.3}
+
+\author[1]{Jameson Nash}
+\author[1]{Jeff Bezanson}
+\author[2]{Kiran Pamnany}
+\affil[1]{Julia Computing, Inc.}
+\affil[2]{Intel Corporation}
+
+\keywords{Julia, JuliaLang, Multithreading, Parallelism, Concurrency, Optimization, Compiler, Runtime VM}
