### 2019-09-09 2c43845
Updates from Overleaf
