+++ paper/paper.tex
@@ -4,7 +4,6 @@
 \documentclass{juliacon}
 \usepackage{hgblistings}
 \usepackage[many]{tcolorbox}
-\usepackage[framemethod=tikz]{mdframed}
 \lstMakeShortInline[language=Julia,columns=fixed]|
 \setcounter{page}{1}
 
