+++ paper/paper.yml
@@ -20,7 +20,6 @@ authors:
 affiliations:
   - name: Julia Computing, Inc.
     index: 1
-affiliations:
   - name: Intel Corporation
     index: 2
 date: 31 August 2019
