### 2019-09-24 #49 @ArchRobison
I had a few comments,  mostly editorial.  Here's a PDF with my markup [10.21105.jcon.00049.pdf](https://github.com/JuliaCon/proceedings-review/files/3644900/10.21105.jcon.00049.pdf)
