### 2019-09-24 #49 @ArchRobison
I had a few comments.  Here's a copy with my markup.
[10.21105.jcon.00049 - ADR comments.pdf](https://github.com/JuliaCon/proceedings-review/files/3644904/10.21105.jcon.00049.-.ADR.comments.pdf)
