### 2019-09-24 #54 @whedon
[REVIEW]: Basic Threading Examples in JuliaLang v1.3
**Submitting author:** @vtjnash (<a href="http://orcid.org/0000-0002-8991-0523">Jameson Nash</a>)
**Repository:** <a href="https://github.com/vtjnash/JuliaCon2019Threading" target ="_blank">https://github.com/vtjnash/JuliaCon2019Threading</a>
