### 2019-10-03 7556fe4
Updates from Overleaf
