+++ paper/paper.tex
@@ -1,4 +1,3 @@
-
 % Based on the JuliaCon proceedings template
 % Layout inspired by the International Journal of Computer Applications template
 \documentclass{juliacon}
@@ -63,9 +62,9 @@
 
 \begin{abstract}
 
-A major distinguishing point of any programming language is how it deals with concurrency. Of course, I'm really interested here in parallelism, and specifically how to get the best throughput out of my application. But we know that actually taking advantage of all cores correctly and efficiently is really hard as a programmer. Here we take a look at how JuliaLang prepares to unleash the full power of your modern CPU's many cores.
+A major distinguishing point of any programming language is how it deals with concurrency. Programmers want to extract the best throughput possible for their applications, but it is well known that taking advantage of all available CPU cores correctly and efficiently is hard. Here, we look at how JuliaLang unleashes the full power of a modern CPU's multiple cores.
 
-Hopefully we won't drive any users mad with frustration as they try to write code to utilize threads. Towards that end, we'll look at how JuliaLang looks to provide a range of modern primitives that are designed to automatically compose effectively, and some of the trade-offs we make to try to simplify the mental model for the user. We'll also take a look at where we think that this development should go towards next.
+One of our key considerations is to reduce the programmer's burden. We will discuss how JuliaLang aims to provide a range of modern primitives that are designed to automatically compose effectively, and some of the trade-offs we make to try to simplify the mental model for the programmer. We'll also briefly discuss our thoughts on future development.
 
 
 \end{abstract}
@@ -73,21 +72,23 @@ Hopefully we won't drive any users mad with frustration as they try to write cod
 \section{Introduction}
 \label{subsub:intro}
 
-A design principle for JuliaLang is to make common tasks easy and difficult tasks possible. This has extended from automatic memory management vs manual memory reuse, to having one type of method with dispatch, to optional type-inference for performance. And now, this includes concurrency. By extending design work that has been present in the language for many years to provide parallelism, we developed a means to add parallelism that preserved the (relative) simplicity of single-threaded for existing code, while allowing new code to benefit from the possibility of threading. This work has been inspired by parallel programming systems like Cilk, Intel Threading Building Blocks (TBB) and Go.
+A design principle for JuliaLang is to make common tasks easy and difficult tasks possible. This is demonstrated in multiple aspects of the language, from automatic memory management vs.\ manual memory reuse, to having one type of method with dispatch, to optional type-inference for performance. We now extend this to concurrency. By extending design work that has been present in the language for many years to provide concurrency, we have developed a means to add parallelism that preserves the (relative) simplicity of single-threaded execution for existing code, while allowing new code to benefit from multi-threaded execution. This work has been inspired by parallel programming systems such as Cilk, Intel Threading Building Blocks (TBB) and Go.
 
-In this paradigm, any piece of a program can be marked for execution in parallel, and a “task” will be started to run that code automatically on an available thread. A dynamic scheduler handles making cache-aware decisions for when to launch the work items. This model of parallelism has many wonderful properties. We see it as somewhat analogous to garbage collection: with GC, you freely allocate objects without worrying about when and how they are freed. With task parallelism, you freely spawn tasks—potentially millions of them—without worrying about where they may eventually run.
[…]
