### 2019-11-27 c3d2c9f
Add pmerge and other updates from Overleaf
