+++ paper/paper.tex
@@ -144,7 +144,7 @@ Notice here how we leverage the do-block syntax to additionally re-use the loop 
 
 This works well for latency bound activities such as a web server. Note that this code only uses one core, although as it performs only thread-local actions, it \emph{could} be run simultaneously on multiple cores without significant change. However, there is also a lot of code that is not written with the expectation of simultaneous access. As such, we want to continue to provide concurrency\footnote{When at least two threads are making progress; a more general form of parallelism that can include time-slicing.} as well as parallelism\footnote{When at least two threads are executing simultaneously}. If we define a thread as a unit of work managed by the runtime system, we can call this $N \mathbin{:} 1$ threading, where the runtime library manages $N$ independent operations and maps them onto one system thread (approximately representing a CPU core). In the new system, we'll let the user now additionally have $M$ CPU cores. This goes beyond the classic $N \mathbin{:} M$ threading model however, as the programmer can specify cooperative affinities where certain tasks won't interrupt each other. We'll call this $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling since we're combining the advantages of single-threaded work queues with multiple cores.
 
-In $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling, we have $N$ units of work mapped onto $M$ CPU threads. Additionally, each of those units of work may be composed of $k$ cooperatively scheduled tasks. This is achieved by pinning the $k$ tasks in a group to one CPU thread, while load balancing units of work across the available cores using the \verb|partr| scheduler, a novel implementation of a parallel depth-first scheduling algorithm.
+In $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling, we have $N$ units of work mapped onto $M$ CPU threads. Additionally, each of those units of work may be composed of $k$ cooperatively scheduled tasks. This is achieved by pinning the $k$ tasks in a group to one CPU thread, while load balancing units of work across the available cores using the \verb|partr| scheduler, a novel implementation of a parallel depth-first scheduling algorithm. This will discussed future in section \ref{subsub:implementation}.
 
 We can extend this model further by factoring out a common factor of $P \mathbin{:} P$ to write this as $N (k \mathbin{:} 1) \mathbin{:} M\ +\ P(k \mathbin{:} 1) \mathbin{:} P$ and derive one further useful use case: the ability to pin one thread to running one task (or task-group). These $P$ tasks could be an over-subscription of the CPU cores, or take away from $M$, or both. Typical uses for this mode of operation would be high-availability tasks (with low latency requirements, but also minimal computation), such as background I/O processing, blocking work pools (foreign library integration), finalizers, or message server queues.
 
@@ -251,10 +251,17 @@ function psort(v::AbstractVector)
     left = fetch(half)::typeof(right)
 # perform the merge on the result
     out = similar(v)
+    merge!(left, right, out)
+    return out
+end
+
+function merge!(left, right, out)
+    ll, lr = length(left), length(right)
+    @assert ll + lr == length(out)
     i, il, ir = 1, 1, 1
-    @inbounds while il <= mid && ir <= (hi - mid)
+    @inbounds while il <= ll && ir <= lr
         l, r = left[il], right[ir]
-        if r < l
+        if isless(r, l)
             out[i] = r
             ir += 1
         else
@@ -263,12 +270,12 @@ function psort(v::AbstractVector)
         end
         i += 1
     end
-    @inbounds while il <= length(left)
+    @inbounds while il <= ll
         out[i] = left[il]
         il += 1
         i += 1
     end
-    @inbounds while ir <= length(right)
+    @inbounds while ir <= lr
         out[i] = right[ir]
         ir += 1
         i += 1
@@ -282,6 +289,8 @@ To see the timing results as we add threads, refer to figure \ref{fig:scalings} 
[…]
