+++ paper/ref.bib
@@ -81,3 +81,40 @@ isbn="978-3-319-20119-1"
  url    = "https://www.mono-project.com/docs/advanced/runtime/docs/coop-suspend/",
  year   = "Apr 16, 2018"
 }
+
+@article{Gavril:1975:MPP:361020.361216,
+ author = {Gavril, F\u{a}nic\u{a}},
+ title = {Merging with Parallel Processors},
+ journal = {Commun. ACM},
+ issue_date = {Oct. 1975},
+ volume = {18},
+ number = {10},
+ month = oct,
+ year = {1975},
+ issn = {0001-0782},
+ pages = {588--591},
+ numpages = {4},
+ url = {http://doi.acm.org/10.1145/361020.361216},
+ doi = {10.1145/361020.361216},
+ acmid = {361216},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {parallel binary insertion, parallel merging, parallel processing},
+} 
+
+@manual{parallel-merge-intel,
+ title  = "A Parallel Stable Sort Using C++ for TBB, Cilk Plus, and OpenMP",
+ author = "Arch D. Robison",
+ url    = "https://software.intel.com/en-us/articles/a-parallel-stable-sort-using-c11-for-tbb-cilk-plus-and-openmp",
+ year   = "Apr 11, 2014"
+}
+
+@book{McCool:2012:SPP:2385466,
+ author = {McCool, Michael and Reinders, James and Robison, Arch},
+ title = {Structured Parallel Programming: Patterns for Efficient Computation},
+ year = {2012},
+ isbn = {9780123914439, 9780124159938},
+ edition = {1st},
+ publisher = {Morgan Kaufmann Publishers Inc.},
+ address = {San Francisco, CA, USA},
+} 
