### 2019-11-27 5f96d18
Typo fixes from Overleaf
