+++ paper/paper.tex
@@ -144,7 +144,7 @@ Notice here how we leverage the do-block syntax to additionally re-use the loop 
 
 This works well for latency bound activities such as a web server. Note that this code only uses one core, although as it performs only thread-local actions, it \emph{could} be run simultaneously on multiple cores without significant change. However, there is also a lot of code that is not written with the expectation of simultaneous access. As such, we want to continue to provide concurrency\footnote{When at least two threads are making progress; a more general form of parallelism that can include time-slicing.} as well as parallelism\footnote{When at least two threads are executing simultaneously}. If we define a thread as a unit of work managed by the runtime system, we can call this $N \mathbin{:} 1$ threading, where the runtime library manages $N$ independent operations and maps them onto one system thread (approximately representing a CPU core). In the new system, we'll let the user now additionally have $M$ CPU cores. This goes beyond the classic $N \mathbin{:} M$ threading model however, as the programmer can specify cooperative affinities where certain tasks won't interrupt each other. We'll call this $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling since we're combining the advantages of single-threaded work queues with multiple cores.
 
-In $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling, we have $N$ units of work mapped onto $M$ CPU threads. Additionally, each of those units of work may be composed of $k$ cooperatively scheduled tasks. This is achieved by pinning the $k$ tasks in a group to one CPU thread, while load balancing units of work across the available cores using the \verb|partr| scheduler, a novel implementation of a parallel depth-first scheduling algorithm. This will discussed future in section \ref{subsub:implementation}.
+In $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling, we have $N$ units of work mapped onto $M$ CPU threads. Additionally, each of those units of work may be composed of $k$ cooperatively scheduled tasks. This is achieved by pinning the $k$ tasks in a group to one CPU thread, while load balancing units of work across the available cores using the \verb|partr| scheduler, a novel implementation of a parallel depth-first scheduling algorithm. This will be discussed further in section \ref{subsub:implementation} on implementation.
 
 We can extend this model further by factoring out a common factor of $P \mathbin{:} P$ to write this as $N (k \mathbin{:} 1) \mathbin{:} M\ +\ P(k \mathbin{:} 1) \mathbin{:} P$ and derive one further useful use case: the ability to pin one thread to running one task (or task-group). These $P$ tasks could be an over-subscription of the CPU cores, or take away from $M$, or both. Typical uses for this mode of operation would be high-availability tasks (with low latency requirements, but also minimal computation), such as background I/O processing, blocking work pools (foreign library integration), finalizers, or message server queues.
 
@@ -290,7 +290,7 @@ While not demonstrated here, \verb|fetch| would also automatically propagate err
 
 Since we are using in-process threads, we could further optimize this to instead mutate the input in-place and to reuse work buffers for additional performance. We have elsewhere tested that and shown the performance improvement is as expected. However, since the scaling improvement was similar between them, we've opted not to include it here.
 
-The algorithm given here is limited in the theoretical scaling capability, since the merge step is not parallelized. On large core counts, that can be important, so please see our supplementary code in \ref{subsub:pmerge}.
+The algorithm given here is limited in the theoretical scaling capability, since the merge step is not parallelized. On large core counts, that can be important, so please see our supplementary code in appendix \ref{subsub:pmerge} for the version with optimal theoretical scaling.
 
[…]
