### 2019-12-22 #54 @vtjnash
Is there anything else I need to do? I think I've adopted and incorporated all of Arch's feedback and made it better.
