### 2020-02-26 41b4a1b
Fix S61_SIEVE code
