+++ paper/paper.tex
@@ -306,37 +306,39 @@ This use of threaded channels is translated from Thomas Hoare's seminal 1978 pap
 \end{figure}
 
 \begin{lstlisting}[language = Julia]
-function S61_SIEVE(n::Integer)
-    buffer = 5 # buffer capacity (configurable)
-    primes = Int[] # result
-    done = Threads.Atomic{Bool}(false) # completion signal
-    sieves = [Channel{Int}(buffer) for i = 1:n]
-    for i in 1:n
+function S61_SIEVE(numPrimes::Integer, nqueue::Int=5)
+    done = Threads.Atomic{Bool}(false)
+    primes = Int[]
+    sieves = [Channel{Int}(nqueue) for i = 1:numPrimes]
+    for i in 1:numPrimes
         Threads.@spawn begin
-            p = take!(sieves[i])
+            sieve = sieves[i]
+            p = take!(sieve)
             push!(primes, p)
             if length(primes) == numPrimes
+                # don't pass it on--we're done now
+                #= TODO: add an atomic write release barrier here =#
                 done[] = true
                 return
             end
-            mp = p # non-prime multiples of p
-            for m in sieves[i]
+            mp = p # mp is a multiple of p
+            for m in sieve
                 while m > mp
                     mp += p
                 end
-                if m < mp # rem(m, p) > 0
+                if m < mp
                     put!(sieves[i + 1], m)
                 end
             end
         end
     end
-
     put!(sieves[1], 2)
     n = 3
     while !done[]
         put!(sieves[1], n)
         n += 2
     end
+    foreach(close, sieves)
     return primes
 end
 \end{lstlisting}
