### 2020-06-08 #54 @wsmoses
@vtjnash Do you need to @sync in the merge sort example before merging or am I missing something.
