### 2020-06-08 #54 @vtjnash
No, that's implied in the `fetch`
