### 2020-06-08 #54 @wsmoses
Ah that makes sense, in any case here's a more full review/set of comments:

This paper presents a mechanism of embedding dynamic task-based parallelism into the Julia language, allowing Julia programs to benefit from parallelism as distinct from existing Julia mechanisms from concurrency. The core of this contribution is a new scheduler with frontend modifications that allow several front-end improvements to make it easier to use this model. The authors also describe several changes to language constructs needed to integrate their framework with arbitrary code, including a demonstration of using external libraries with their framework.

The authors present several demonstrations of programs (merge sort, number sieve, and a prefix scan), that could benefit from the task-parallelism that they now permit. The additional syntax, while demonstrated in code blocks, isn't explained well for new users hoping to take advantage of it (e.g. no description of @sync/fetch, ambiguity about the guarantees provided by @spawn).

The authors allude to several limitations in existing Julia threading primitives. There isn't a good description of why the existing threading primitives couldn't be used on the demonstration codes (though I can see why this would be difficult). It would also be useful to do a performance comparison between the new runtime and old systems on some mutually compatible code. This comparison, along with perhaps showing the serial runtime [e.g. without parallel constructs rather than 1-core with parallel constructs] would be helpful to demonstrate that the new runtime doesn't provide significant overhead.

It would also be desirable to provide more quantifiable justification of their performance on FFTW than simply claiming it to be competitive (which is somewhat unclear). The authors should also include additional reference to related work in the field (e.g. existing parallel runtimes, including those outside of Julia) and describe the similarities and differences between prior art and the runtime described in this paper.

As a minor point (and I'm not the most familiar with the types of papers in this venue so take this with a grain of salt): some of the wording e.g. "but on to even more fun stuff..." could be tightened up for a professional audience.

Overall this paper represents a significant contribution to the Julia community in improving the experience of writing high performance parallel code. There are some additional changes to the writeup that would aid in fully demonstrating this contribution.
