### 2020-08-13 e85bf6b
More changes by Jameson and Kiran to address reviewer comments
