+++ paper/paper.tex
@@ -1,4 +1,4 @@
- % Based on the JuliaCon proceedings template
+% Based on the JuliaCon proceedings template
 % Layout inspired by the International Journal of Computer Applications template
 \documentclass{juliacon}
 \usepackage{hgblistings}
@@ -149,15 +149,9 @@ In $N (k \mathbin{:} 1) \mathbin{:} M$ scheduling, we have $N$ units of work map
 
 We can extend this model further by factoring out a common factor of $P \mathbin{:} P$ to write this as $N (k \mathbin{:} 1) \mathbin{:} M\ +\ P(k \mathbin{:} 1) \mathbin{:} P$ and derive one further useful use case: the ability to pin one thread to running one task (or task-group). These $P$ tasks could be an over-subscription of the CPU cores, or take away from $M$, or both. Typical uses for this mode of operation would be high-availability tasks (with low latency requirements, but also minimal computation), such as background I/O processing, blocking work pools (foreign library integration), finalizers, or message server queues.
 
-The newest work here is the implementation of a thread-aware scheduler. For a couple years, JuliaLang has been able to perform simple loop parallelism with numerous caveats including a static work schedule and minimal to none inter-thread communication. This had been labelled ``experimental,'' while we worked out the details and shook out the bugs. The recent work to generalize that functionality has allowed us to rewrite that older static capability on top of our new runtime, demonstrating the power and flexibility of the new system. While the old system let us test out the thread-related bugs in the runtime, it had some significant shortcomings. We knew those would need to be addressed before we could move beyond the ``experimental'' label.
+For a couple years, JuliaLang has been able to perform simple loop parallelism with the `\verb|@threads for|' macro, roughly analogous to OpenMP's `\verb|#pragma omp parallel for schedule(static)|' without support for reductions. This had been labelled ``experimental'' while we focused on making JuliaLang's runtime re-entrant and threadsafe and clarified requirements for the final parallelism capabilities. The experimental threading infrastructure had no scheduler, could not interact with regular tasks or do I/O, and parallel loops could not be nested. This made it nearly impossible to write many common algorithms or use a large portion of the language while running a threaded region of code.
 
-Those limitations included:
-\begin{itemize}
-\item it had no scheduler, and so couldn't interact with regular tasks or do I/O
-\item the runtime was full of race conditions
-\item the `\verb|@threads|' loops could not be nested
-\item parallel code couldn't run outside the `\verb|@threads|' loop
-\end{itemize}
+The new threading runtime addresses all these shortcomings. Furthermore, in addition to the new parallelism constructs that have been introduced, the previous loop parallelism capability has been rewritten on top of this runtime, demonstrating its power and flexibility.
 
 \begin{calloutbox}[Running Julia with Threads]
 In the examples below, we will be using JuliaLang v1.3 launched with multiple threads.
@@ -187,7 +181,7 @@ julia> Threads.threadid()
 \subsection{Stochastic Ordering}
 \label{subsub:pprint}
 
-One of the more visual ways to show we have threads working is to show the scheduler picking up work in semi-random, interleaving, orders. Previous versions of Julia already had an `\verb|@threads for|' macro which would split a range and run a portion on each thread with a static schedule. So in the range below, thread 1 would run items 1 and 2, thread 2 would run items 3 and 4, and so on. Now these threads supports doing I/O too with that same schedule.
[…]
