### 2021-01-14 #54 @vchuravy
>  don't have an archival DOI, but don't think it'll be useful for this.

Can you create one with the example code being used?
