+++ code/prefix_scan_sum.jl
@@ -1,0 +1,25 @@
+# This code license is MIT: https://github.com/vtjnash/JuliaCon2019Threading/blob/master/LICENSE
+
+using Base.Threads: @threads
+
+function prefix_threads!(⊕, y::AbstractVector)
+    l = length(y)
+    k = ceil(Int, log2(l))
+    # do reduce phase
+    for j = 1:k
+        @threads for i = 2^j:2^j:min(l, 2^k)
+            @inbounds y[i] = y[i - 2^(j - 1)] ⊕ y[i]
+        end
+    end
+    # do expand phase
+    for j = (k - 1):-1:1
+        @threads for i = 3*2^(j - 1):2^j:min(l, 2^k)
+            @inbounds y[i] = y[i - 2^(j - 1)] ⊕ y[i]
+        end
+    end
+    return y
+end
+
+# Example:
+#   A = fill(1, 500_000)
+#   prefix_threads!(+, A)
