+++ code/psort.jl
@@ -1,0 +1,83 @@
+# This code license is MIT: https://github.com/vtjnash/JuliaCon2019Threading/blob/master/LICENSE
+
+using Base.Threads: @spawn
+
+function psort(v::AbstractVector)
+    hi = length(v)
+    if hi < 100_000 # below some cutoff, run in serial
+        return sort(v, alg = MergeSort)
+    end
+    # split the range and sort the halves in parallel recursively
+    mid = (1 + hi) >>> 1
+    half = @spawn psort(view(v, 1:mid))
+    right = psort(view(v, (mid + 1):hi))
+    left = fetch(half)::typeof(right)
+    # perform the merge on the result
+    out = similar(v)
+    pmerge!(out, left, right)
+    return out
+end
+
+function merge!(out, left, right)
+    ll, lr = length(left), length(right)
+    @assert ll + lr == length(out)
+    i, il, ir = 1, 1, 1
+    @inbounds while il <= ll && ir <= lr
+        l, r = left[il], right[ir]
+        if isless(r, l)
+            out[i] = r
+            ir += 1
+        else
+            out[i] = l
+            il += 1
+        end
+        i += 1
+    end
+    @inbounds while il <= ll
+        out[i] = left[il]
+        il += 1
+        i += 1
+    end
+    @inbounds while ir <= lr
+        out[i] = right[ir]
+        ir += 1
+        i += 1
+    end
+    return out
+end
+
+function pmerge!(out, left, right)
+    ll, lr = length(left), length(right)
+    @assert ll + lr == length(out)
+    if length(out) < 100_000
+        # below some threshold, just do the merge
+        merge!(out, left, right)
+    else
+        # split the larger chunk in half, then binary search the
+        # smaller half to split it
+        if ll > lr
+            jl = ll ÷ 2
+            # stable sort: find the last entry in right
+            # strictly smaller than l
+            jr = searchsortedfirst(right, left[jl]) - 1
+        else
+            jr = lr ÷ 2
+            # stable sort: find the last entry in left not bigger
+            # than r
+            jl = searchsortedlast(left, right[jr])
+        end
+        @sync begin
+            let left = view(left, 1:jl),
+                right = view(right, 1:jr),
+                out = view(out, 1:(jl + jr))
+                @spawn pmerge!(out, left, right)
+            end
+            let left = view(left, (jl + 1):ll),
+                right = view(right, (jr + 1):lr),
+                out = view(out, (jl + jr + 1):length(out))
+                pmerge!(out, left, right)
+           end
+        end
+    end
+    nothing
+end
