+++ README.md
@@ -1,1 +1,3 @@
+[![DOI](https://zenodo.org/badge/206995495.svg)](https://zenodo.org/badge/latestdoi/206995495)
+
 See [Threading_JuliaCon2019.pdf](Threading_JuliaCon2019.pdf)
